Prototype of correct reachability management for objects with native counterparts.
As described in https://stackoverflow.com/questions/24376768/can-java-finalize-an-object-when-it-is-still-in-scope/24380219#24380219
it's not guaranteed that objects do not reach finalization queue even if they seemingly referred by local variables.
So as was seen in https://github.com/JetBrains/compose-jb/issues/124 we could invoke finalizer when long running native
method using resource is still in progress, and so - crash the VM.

In this fix we instead pass object handle to the JNI method instead, and it guarantees that object is reachable and not finalized.

This fix is POC, and generally scheme like that shall be applied to all objects with native counterparts in Skija.